Factorisation et divers modif

+++ Librairies/Controller/TravauxController.php
@@ -2,21 +2,21 @@
 /*
 Author: fpodev (fpodev@gmx.fr)
 TravauxController.php (c) 2020
-Desc: description
+Desc: script de controle pour les travaux.
 Created:  2020-04-20T08:32:18.988Z
 Modified: !date!
 */
 namespace App\Controller;
 
-use App\View\Render;
 use App\Objet\Travaux;
 use App\model\LieuModel;
 use App\model\UserModel;
-use App\Objet\ConnexionDb;
+use App\ConnexionBDD\ConnexionDb;
 use App\model\SecteurModel;
 use App\model\TravauxModel;
 use App\model\BatimentModel;
 use App\model\MaterielModel;
+use App\Controller\VueController;
 use App\Controller\UserController;
 
 class TravauxController {
@@ -30,7 +30,7 @@ class TravauxController {
     public function __construct(){
         $db = ConnexionDb::getPDO();
         $this->userCtrl = new UserController();
-        $this->render = new Render();  
+        $this->render = new VueController;  
         $this->travaux = new TravauxModel($db);   
         $this->ville = new LieuModel($db);         
         $this->secteur = new SecteurModel($db);
@@ -39,18 +39,19 @@ class TravauxController {
         $this->batiment = new BatimentModel($db);
         $this->date = new \DateTime();
     }
-    
+    /*affiche les travaux programmés pour un niveau 2*/
     public function travauxTech(){                                                    
             $title = "Liste de mes travaux";                
 
             $this->render->view('TravauxTech',['title' => $title]);              
     }
+    /*$etat permet de gérer les différents droit sur la page afiché*/
+    /*$param permet de chercher dans la BDD les travaux planifié ou non*/
     public function travauxList(){
         if(stristr((urldecode($_SERVER['REQUEST_URI'])), 'travaux-planifiés')){
             $title = "Liste des travaux planifiés";                
             $param = 'NOT NULL';
-            $etat ='1';
-            
+            $etat ='1';            
         }                       
         else
         {
@@ -64,15 +65,31 @@ class TravauxController {
 
     }
     public function uniqueTravaux($id){
-            $_SESSION['lieuId'] = "null";
-            $id = $id;
-            $title = "Planification des travaux demandés";
-            $value = $this->travaux->uniqueTravaux($id);                    
-            $techList = $this->user->listTech($_SESSION['lieuId']);
-           
-            $this->render->view('PlanifTravaux', ['title' => $title,'tvx' => $value, 'tech' => $techList]);
-    }
-    public function addTravaux(){                   
+            $id = $id;        
+                                                 
+            $value = $this->verif($id);
+
+            if($value != FALSE){                
+                foreach($value as $idLieu)  
+                         
+                $techList = $this->user->listTech($idLieu['nLieu']);  
+
+                if(stristr($_SERVER['REQUEST_URI'], 'edit-travaux=')){
+                    $title = "Modifier la demande de travaux";                
+                }
+                elseif(stristr($_SERVER['REQUEST_URI'], 'voir=')){          
+                    $title = "Voici la demande de travaux";              
+                }
+                else{
+                    $title = "Planification de la demande de travaux";                
+                }
+                $this->render->view('PlanifTravaux', ['title' => $title,'tvx' => $value, 'tech' => $techList]); 
+                }                                      
+    }          
+    public function addTravaux(){  
+        $description = htmlspecialchars($_POST['descriptions']);
+        $detail = htmlspecialchars($_POST['detail']);
+        $urgence = htmlspecialchars($_POST['urgence']);       
             $travaux = new Travaux(                     
                 [               
                     'id_lieu' => $_POST['id_lieu'],
@@ -80,27 +97,27 @@ class TravauxController {
                     'id_secteur' => $_POST['id_secteur'],
                     'id_materiel' => $_POST['id_materiel'],
                     'id_demandeur' => $_POST['id_demandeur'],                
-                    'descriptions' => $_POST['descriptions'],
-                    'detail' => $_POST['detail'],
-                    'urgence' => $_POST['urgence'], 
-                    'descriptions' => $_POST['descriptions'],
-                    'detail' => $_POST['detail']                                                              
+                    'descriptions' => $description,                    
+                    'urgence' => (int)$urgence,   
+                    'detail' => $detail,                                                                          
                 ]            
-                );
-                              
+                );                              
         if($travaux->validDemande()) 
         {            
             $this->travaux->save($travaux);
+            $this->travauxList();
         }
         else
         {
             $erreurs = $travaux->erreurs();                     
-            
-            $this->render->view('DemandeTvx', ['tvx' => $erreurs]); 
-        } 
-         $this->userCtrl->home();        
+            $title = 'Formulaire demande de travaux';
+            $this->render->view('DemandeTvx', ['tvx' => $erreurs, 'title' => $title]); 
+        }             
     }
     public function planifTravaux(){
+        $description = htmlspecialchars($_POST['descriptions']);
+        $detail = htmlspecialchars($_POST['detail']); 
+        
             $travaux = new Travaux(                     
                 [        
                     'descriptions' => $_POST['descriptions'],
@@ -110,13 +127,14 @@ class TravauxController {
                     'externe' => $_POST['externe'],                              
                 ]            
                 ); 
-                 
+                
         if(isset($_POST['id']))  
         {
             $travaux->setId($_POST['id']);
         }        
         if($travaux->validPlanif()) 
         { 
+            
             $this->travaux->save($travaux);
             $this->travauxList();
         }
@@ -126,89 +144,41 @@ class TravauxController {
             
             $this->render->view('PlanifTravaux', ['tvx' => $erreurs]);
         }
-    } 
-    public function startTravaux($donnee){        
-        $travaux = new Travaux(                     
-            [                                                                                                              
-                'date_debut' => $this->date->format('d-m-Y H:i:s'),                                                                 
-            ]            
-            ); 
-               
-        if(isset($donnee->id))  
-        {
-            $travaux->setId($donnee->id);
-        }      
-        if($travaux->validStart()) 
-        { 
-            $this->travaux->save($travaux);                      
-        }
-        else
-        {
-            $erreurs = $travaux->erreurs();                     
-            
-            $this->render->view('DemandeTvx', ['tvx' => $erreurs]);
-        }
     }  
-    public function valid($travaux){
-            
-    }
-    public function closeTravaux(){        
-        $travaux = new Travaux(                     
-            [                                                                                                        
-               'date_fin'=> $this->date->format('Y-m-d H:i:s'),                                         
-            ]            
-            );                 
-        if(isset($_POST['id']))  
-        {
-            $travaux->setId($_POST['id']);
-        }
-        
-        if($travaux->validClose()) 
-        { 
-            $this->travaux->save($travaux);
-            
-            $this->travauxList();
-        }
-        else
-        {
-            $erreurs = $travaux->erreurs();                     
-            
-            $this->render->view('DemandeTvx', ['tvx' => $erreurs]);
-        } 
-    }         
+           
     public function travauxPage($id = null){
         if($id == null){            
             $id = $_SESSION['lieuId'];            
-        }            
+        }                       
         if(preg_match("#[0-9]#", $id))
             {                                                   
                 if(stristr($_SERVER['REQUEST_URI'], 'Demande-Travaux'))                
                 {
                     $title = 'bâtiment';  
-                    $ville = $this->ville->uniqueLieu($id);   
+                    $ville = $this->ville->uniqueLieu(htmlspecialchars($id));  
                     $_SESSION['lieuVue'] = $ville->nom();                
-                    $valueList = $this->batiment->batimentList($id);                    
+                    $valueList = $this->batiment->batimentList(htmlspecialchars($id));                    
                 }                                
                 elseif(stristr((urldecode($_SERVER['REQUEST_URI'])), 'Bâtiment', true))
                 {  
                     $title = 'secteur';  
-                    $valueList = $this->secteur->secteurList($id);  
-                    $batiment = $this->batiment->uniqueBatiment($id);
+                    $batiment = $this->batiment->uniqueBatiment(htmlspecialchars($id));
+                    $valueList = $this->secteur->secteurList(htmlspecialchars($id));                   
                     $_SESSION['batiment'] = $batiment->nom(); 
                     $_SESSION['batimentId'] = $batiment->id();
                 }
                 elseif(stristr($_SERVER['REQUEST_URI'], 'Secteur'))
                 {
                     $title = 'matériel'; 
-                    $valueList = $this->materiel->materielList($id);
-                    $secteur = $this->secteur->uniqueSecteur($id);
+                    $valueList = $this->materiel->materielList(htmlspecialchars($id)); 
+                    $secteur = $this->secteur->uniqueSecteur(htmlspecialchars($id));
                     $_SESSION['secteur'] = $secteur->nom();  
                     $_SESSION['secteurId'] = $secteur->id();                                       
                 }                                                                                       
                 $this->render->view('ActifTvx', ['value' => $valueList, 'title' => $title]); 
             }
             else{
-                echo 'erreur 404';
+                $this->render->view('404');
             }            
     }
     public function formulaireTravaux($id = null){
@@ -218,19 +188,41 @@ class TravauxController {
         if(preg_match("#[0-9]#", $id))
         {
             $title = 'Formulaire demande de travaux';
-            $materiel = $this->materiel->uniqueMateriel($id);
+            $materiel = $this->materiel->uniqueMateriel(htmlspecialchars($id));
             $_SESSION['materiel'] = $materiel->nom();
             $_SESSION['materielId'] = $materiel->id();
 
             $this->render->view('DemandeTvx', ['title' => $title]);
         }
+        else{
+            $this->render->view('404');
+        }
     }      
 
-   public function deleteTravaux($id){
-       
-        $this->travaux->delete($id);     
-
-        $this->travauxList();
-                
+   public function deleteTravaux($id){      
+        $travaux = $this->verif($id);
+        if($travaux == true){           
+            $this->travaux->delete($id);   
+            $this->travauxList();
+        }                            
    }
+   /*verification que l'élement demandé via un $id est bien un nombre
+    ** et existe bien dans la BDD avant de retourné ses valeurs*/
+   public function verif($id){    
+    $idValid =  htmlspecialchars($id);           
+                                  
+    if(preg_match("#[0-9]#", $idValid))
+    {     
+       $travaux = $this->travaux->uniqueTravaux(htmlspecialchars($id));          
+        if($travaux != false){
+            return $travaux;
+        } 
+        else{
+            $this->render->view('404');
+        }                            
+    } 
+    else{                             
+        $this->render->view('404');                          
+    }           
+}    
 }
